+++ src/types/index.ts
@@ -1,1 +1,15 @@
 export type Direction = 'up' | 'right' | 'down' | 'left';
+
+export type FormattedDeviceType = {
+  deviceId: string;
+  label: string;
+};
+
+export type DeviceType = MediaDeviceInfo | FormattedDeviceType;
+
+export type SelectedDeviceType = string | null;
+
+export type DeviceTypeContext = {
+  devices: DeviceType[];
+  selectedDevice: SelectedDeviceType;
+};